Move the display ingest models into their own namespace
