+++ api/ingests_api/src/main/scala/uk/ac/wellcome/platform/storage/ingests/api/IngestsApi.scala
@@ -6,7 +6,7 @@ import akka.http.scaladsl.server._
 import akka.http.scaladsl.server.Directives._
 import de.heikoseeberger.akkahttpcirce.ErrorAccumulatingCirceSupport._
 import uk.ac.wellcome.json.JsonUtil._
-import uk.ac.wellcome.platform.archive.display.RequestDisplayIngest
+import uk.ac.wellcome.platform.archive.display.ingests.RequestDisplayIngest
 import uk.ac.wellcome.platform.storage.ingests.api.responses.{
   CreateIngest,
   LookupIngest
