+++ display/src/main/scala/uk/ac/wellcome/platform/archive/display/ingests/DisplayBag.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.display
+package uk.ac.wellcome.platform.archive.display.ingests
 
 import io.circe.generic.extras.JsonKey
