+++ display/src/main/scala/uk/ac/wellcome/platform/archive/display/ingests/DisplayCallback.scala
@@ -1,4 +1,5 @@
-package uk.ac.wellcome.platform.archive.display
+package uk.ac.wellcome.platform.archive.display.ingests
+
 import io.circe.generic.extras.JsonKey
 import uk.ac.wellcome.platform.archive.common.ingests.models.Callback
